#25(2) 파일 모듈화 적용 (Board / Account)

+++ src/modules/board/faq/command/delete-faq.handler.ts
@@ -1,36 +1,39 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { CommandHandler, EventBus, ICommandHandler } from '@nestjs/cqrs';
 import { DeleteFaqCommand } from './delete-faq.command';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Faq } from '../entities/faq';
 import { Board } from '../../entities/board';
-import { BoardFile } from '../../../file/entities/board_file';
-import { FileDeleteEvent } from '../../../file/event/file-delete-event';
+import { BoardFile } from '../../../file/entities/board-file';
+import { FilesDeleteEvent } from '../../../file/event/files-delete-event';
+import { BoardFileDb } from '../../board-file-db';
+import { ConvertException } from '../../../../common/utils/convert-exception';
 
 /**
- * FAQ 삭제 시, 커맨드를 처리하는 커맨드 핸들러
+ * FAQ 삭제용 커맨드 핸들러
  */
-
 @Injectable()
 @CommandHandler(DeleteFaqCommand)
 export class DeleteFaqHandler implements ICommandHandler<DeleteFaqCommand> {
   constructor(
-    @InjectRepository(Faq)
-    private faqRepository: Repository<Faq>,
-
-    @InjectRepository(Board)
-    private boardRepository: Repository<Board>,
-
-    @InjectRepository(BoardFile)
-    private fileRepository: Repository<BoardFile>,
-
+    @InjectRepository(Faq) private faqRepository: Repository<Faq>,
+    @InjectRepository(Board) private boardRepository: Repository<Board>,
+    @InjectRepository(BoardFile) private fileRepository: Repository<BoardFile>,
+    @Inject('faqFile') private boardFileDb: BoardFileDb,
+    @Inject(ConvertException) private convertException: ConvertException,
     private eventBus: EventBus,
   ) {}
 
+  /**
+   * FAQ 삭제 메소드
+   * @param command : FAQ 삭제에 필요한 파라미터
+   * @returns : DB처리 실패 시 에러 메시지 반환 / 삭제 성공 시 완료 메시지 반환
+   */
   async execute(command: DeleteFaqCommand) {
     const { faqId, role, accountId } = command;
 
+    // TODO : 권한 정보 데코레이터 적용시 확인 후, 삭제 예정
     if (role !== '본사 관리자') {
       throw new BadRequestException('본사 관리자만 접근 가능합니다.');
     }
@@ -38,39 +41,39 @@ export class DeleteFaqHandler implements ICommandHandler<DeleteFaqCommand> {
     const faq = await this.faqRepository.findOneBy({ faqId: faqId });
 
     if (!faq) {
-      throw new NotFoundException('존재하지 않는 FAQ입니다.');
+      return this.convertException.notFoundError('FAQ', 404);
     }
 
+    // TODO : 유저 정보 데코레이터 적용시 확인 후, 삭제 예정
     if (accountId != faq.boardId.accountId) {
       throw new BadRequestException('작성자만 삭제가 가능합니다.');
     }
 
     const board = await this.boardRepository.findOneBy({ boardId: faq.boardId.boardId });
 
-    const files = await this.fileRepository.findBy({ boardId: board.boardId });
+    if (!board) {
+      return this.convertException.notFoundError('게시글', 404);
+    }
 
-    // 파일 삭제 이벤트 처리
-    this.eventBus.publish(new FileDeleteEvent(board.boardId, files));
+    const boardFiles = await this.fileRepository.findBy({ boardId: board.boardId });
 
-    // board_file db 삭제
-    files.map((file) => {
-      this.fileRepository.softDelete({ boardFileId: file.boardFileId });
-    });
+    if (boardFiles.length !== 0) {
+      // 파일 삭제 이벤트 처리
+      this.eventBus.publish(new FilesDeleteEvent(board.boardId, this.boardFileDb));
+    }
 
-    // faq db 삭제
     try {
       await this.faqRepository.delete(faq);
     } catch (err) {
-      console.log(err);
+      return this.convertException.CommonError(500);
     }
 
-    // board db 삭제 (fk)
     try {
       await this.boardRepository.softDelete({ boardId: board.boardId });
     } catch (err) {
-      console.log(err);
+      return this.convertException.CommonError(500);
     }
 
-    return 'FAQ 삭제 성공';
+    return '삭제가 완료 되었습니다.';
   }
 }
